[nfcd-binder-plugin] Virtualize binder API. JB#63598
Single API is never enough for Google.

An alternative approach would be to have a separate plugin for each
variation of the binder NFC API but at this point it seems feasible
to support multiple backends inside a single plugin, to reduce the
number of packages involved in making the whole thing work.